+++ AmazonChimeSDK/AmazonChimeSDK/internal/audio/protocols/AudioClientProtocol.swift
@@ -23,6 +23,20 @@ import Foundation
                       callKitEnabled: Bool,
                       appInfo: AppInfo!) -> audio_client_status_t
 
+    // swiftlint:disable function_parameter_count variable_name
+    func startSession(_ host: String!,
+                      basePort port: Int,
+                      callId: String!,
+                      profileId: String!,
+                      microphoneMute mic_mute: Bool,
+                      speakerMute spk_mute: Bool,
+                      isPresenter presenter: Bool,
+                      sessionToken tokenString: String!,
+                      audioWsUrl: String!,
+                      callKitEnabled: Bool,
+                      appInfo: AppInfo!,
+                      audioMode: AudioModeInternal) -> audio_client_status_t
+
     func startSession(_ host: String!,
                       basePort port: Int,
                       callId: String!,
@@ -55,9 +69,9 @@ import Foundation
     func isBliteNSSelected() -> Bool
 
     func setBliteNSSelected(_ bliteSelected: Bool) -> Int
-    
+
     func endOnHold()
-    
+
     var delegate: AudioClientDelegate! { get set }
 }